requirements.txt
boto3==1.6.1
botocore==1.9.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
docutils==0.14
facebook-sdk==2.0.0
Flask==0.12.2
futures==3.2.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
python-dateutil==2.6.1
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
yattag==1.10.0
